# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Sports Management API
  description: |
    Comprehensive API for managing sports leagues, referee assignments, games, teams, and budgets.
    
    ## Authentication
    All endpoints (except `/health` and `/auth/login`) require Bearer authentication.
    Include the JWT token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Login endpoint: 5 requests per minute
    - Registration endpoint: 3 requests per hour
    - Password reset: 3 requests per hour
    - Assignment operations: 100 requests per minute
    - General endpoints: 1000 requests per hour
    
    ## Error Handling
    All error responses follow a consistent format:
    ```json
    {
      "error": "Error message",
      "message": "Detailed description",
      "code": "ERROR_CODE"
    }
    ```
    
    ## Data Models
    The API follows a hierarchical structure:
    - Organizations contain Leagues
    - Leagues contain Teams
    - Games reference Teams and are assigned Referees
    - Budgets track financial aspects
    
  version: 1.0.0
  contact:
    name: Sports Management API Support
    email: support@sportsmanagement.app
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.sportsmanagement.app
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    League:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization:
          type: string
          description: Organization name
        age_group:
          type: string
          description: Age group (e.g., U12, U14)
        gender:
          type: string
          enum: [Boys, Girls, Mixed]
        division:
          type: string
          description: Division level
        season:
          type: string
          description: Season identifier
        level:
          type: string
          enum: [Recreational, Competitive, Elite]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - organization
        - age_group
        - gender
        - division
        - season
        - level
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error description
        code:
          type: string
          description: Error code for programmatic handling
...
security:
  - bearerAuth: []
paths:
...
  /leagues:
    get:
      tags:
        - Leagues
      summary: Get all leagues
      description: Retrieve all leagues with optional filtering
      parameters:
        - name: organization
          in: query
          description: Filter by organization
          schema:
            type: string
        - name: age_group
          in: query
          description: Filter by age group
          schema:
            type: string
        - name: gender
          in: query
          description: Filter by gender
          schema:
            type: string
            enum: [Boys, Girls, Mixed]
        - name: division
          in: query
          description: Filter by division
          schema:
            type: string
        - name: season
          in: query
          description: Filter by season
          schema:
            type: string
        - name: level
          in: query
          description: Filter by level
          schema:
            type: string
            enum: [Recreational, Competitive, Elite]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Leagues retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      tags:
        - Leagues
      summary: Create new league
      description: Create a new league (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/League'
            example:
              organization: "Metro Soccer Association"
              age_group: "U12"
              gender: "Boys"
              division: "Division 1"
              season: "Fall 2024"
              level: "Competitive"
      responses:
        '201':
          description: League created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  league:
                    $ref: '#/components/schemas/League'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /leagues/{id}:
    get:
      tags:
        - Leagues
      summary: Get specific league
      description: Retrieve a specific league by ID
      parameters:
        - name: id
          in: path
          required: true
          description: League ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: League retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  league:
                    $ref: '#/components/schemas/League'
        '404':
          description: League not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Leagues
      summary: Update league
      description: Update an existing league (admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: League ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/League'
      responses:
        '200':
          description: League updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  league:
                    $ref: '#/components/schemas/League'
        '404':
          description: League not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Leagues
      summary: Delete league
      description: Delete a league (admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: League ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: League deleted successfully
        '404':
          description: League not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Teams
...